devcontainer: Add dev container Dockerfile
This should include all the non-Node dev tools required for at least the
hello world app. We will add more to this as we find more dev packages
that we need to install.

Also include a `devcontainer.json` file for the VS Code
`remote-containers` extension.
https://marketplace.visualstudio.com/items?itemName=ms-vscode-remote.remote-containers

Since there are npm packages avaialable for most of our dev tools
(namely `ether.js` and `hardhat`), we can work mostly using the Node
project workflow, so this is nowhere as essential as I thought it would
be. Nevertheless, it's nice for all of us to work on the "same OS" and
same versions of Docker and same versions of Vyper and same versions of
node.