Containerize components for development
Add a Dockerfile for bringing up the client in development mode.
The purpose of this is mainly to allow us to use `docker-compose` to
bring (most of) the app up in one command.

Add docker-compose file. Currently, this is not much use, since we
haven't set up our Express server or Mongodb yet. So all it does is
bring up the client development container. We will update this once
we have the other components set up.