docker-compose: Don't serve React in development
We previously encountered
`Error: ENOENT: no such file or directory, stat /var/app/backend/build/index.html`
when starting the backend. However, we do not intend for the backend
Express server to serve our frontend artifacts when running in dev mode.

Update docker-compose file to indicate that we are starting in dev mode
and have the Express server not attempt to serve React artifacts in
development (it is served by another container in development).